new addRow() method that takes column values, not cells
We introduce a new method addRow(Iterable<Object> values) in Table
that allows to create a new row just by providing the values that
should be written to the individual columns. This makes it easier
for the user because they no longer have to wrap each value into a
TableCell instance. The TableCell is for now automatically created
within the Table.

This is in preparation for separating the values that should be
written into the table from their appearance. Currently, a TableCell
contains both, the value itself as well as the logic how to format
it.