Extend TableCell to be reusable as renderer for multiple values
We introduce a new method setValue() for the TableCell interface, that
changes the value that is rendered. This enables us to reuse a table
cell and simplifies the TableModel since we only need to define the
TableCell once and each row can only contain the values, without the
need to configure their appearance. If a TableRow contains a cell,
this cell is used instead, so that we stay backwards compatible.