Add a new class that encapsulates lengths
PDFBox uses an internal unit for measuring lengths that is 1/72 of
an inch. If one designs forms or other documents, it can be helpful
to be able to define lengths in other units like mm or inchs. For
this a new class Length is added that allows to convert across
different measurement units.
At a later point in time, we might consider using Length within
unbox-pdf in all places where we keep lengths and calculate with
them.